Models restructured
The Models were not optimally designed before. So, I fixed them up and
this should improve EF's functionality along with ease of programming
against the database.

The awkward partial views were removed also. Note: partial views will be
rendered by using the @Html.Action() as briefly described in
Authors/Index view.

Prettying up things to do next, along with seeding the database using
Migration so that the project can be tested in the browser as other
improvements continue to come.