Enhancements
A Beer object and its attached Journal object can now be added by the
user.

The Author profile got spruced up a bit and now has a good format for
the three most recent beer and coffee journals.

Required range of Rating for a Journal was removed, getting an int from
a radio button is not possible.

The Abv and IBU fields had required validation when testing even though
they should not be validated as required. That still needs looking into.